LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-030-expected.txt


PASS .test 1
PASS .test 2
FAIL .test 3 assert_equals:
<div class="test small-contents overflow-auto cis-width" data-expected-client-width="100" data-expected-client-height="0" data-expected-scroll-width="100" data-expected-scroll-height="20"></div>
clientWidth expected 100 but got 85
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
PASS .test 9
PASS .test 10
PASS .test 11
PASS .test 12
PASS .test 13
PASS .test 14
FAIL .test 15 assert_equals:
<div class="test big-contents overflow-auto cis-width" data-expected-client-width="100" data-expected-client-height="0" data-expected-scroll-width="400" data-expected-scroll-height="200"></div>
clientWidth expected 100 but got 85
FAIL .test 16 assert_equals:
<div class="test big-contents overflow-auto cis-both" data-expected-client-width="100" data-expected-client-height="50" data-expected-scroll-width="400" data-expected-scroll-height="200"></div>
clientWidth expected 100 but got 85
PASS .test 17
PASS .test 18
PASS .test 19
PASS .test 20
PASS .test 21
PASS .test 22
PASS .test 23
PASS .test 24
